+++ onnx/inliner/inliner.cc
@@ -431,7 +431,7 @@ void ConvertVersion(ModelProto& model, const NodeProto& call_node, FunctionProto
 
   RepeatedNodeProto& function_nodes = *function.mutable_node();
   RepeatedNodeProto& nodes = *graph.mutable_node();
-  nodes.Reserve(function_nodes.size() + used_vars.size());
+  nodes.Reserve(static_cast<int>(function_nodes.size() + used_vars.size()));
 
   auto* inputs = graph.mutable_input();
   for (const auto& var : used_vars) {
@@ -540,7 +540,7 @@ struct InlinerImpl {
     }
     if (schema_registry != nullptr) {
       int64_t domain_version = GetDomainVersion(model, domain);
-      const auto* const op_schema = schema_registry->GetSchema(node.op_type(), domain_version, domain);
+      const auto* const op_schema = schema_registry->GetSchema(node.op_type(), static_cast<int>(domain_version), domain);
 
       if (op_schema == nullptr) {
         // If the schema is not found, we cannot inline the function.
@@ -548,7 +548,7 @@ struct InlinerImpl {
       }
 
       if (op_schema->HasFunction()) {
-        const FunctionProto* function_ptr = op_schema->GetFunction(domain_version, false);
+        const FunctionProto* function_ptr = op_schema->GetFunction(static_cast<int>(domain_version), false);
         if (function_ptr != nullptr) {
           callee = *function_ptr;
           target_version = kNoConversion;
@@ -565,7 +565,7 @@ struct InlinerImpl {
         }
         ONNX_NAMESPACE::FunctionBodyBuildContextImpl function_body_ctx(node, input_types);
         target_version = kNoConversion;
-        return op_schema->BuildContextDependentFunction(function_body_ctx, callee, domain_version);
+        return op_schema->BuildContextDependentFunction(function_body_ctx, callee, static_cast<int>(domain_version));
       }
     }
     return false;
@@ -590,7 +590,7 @@ struct InlinerImpl {
         // Rename variable names in callee
         InliningRenamer::Rename(node, callee, "__" + std::to_string(++(this->inline_count)), this->name_generator);
         if (target_version != kNoConversion) {
-          ConvertVersion(model, node, callee, target_version);
+          ConvertVersion(model, node, callee, static_cast<int>(target_version));
         }
         std::unordered_set<std::string> actual_parameters;
         for (const auto& x : node.input())
